+++ src/main/java/com/woowacamp/soolsool/core/member/code/MemberErrorCode.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.member.exception;
+package com.woowacamp.soolsool.core.member.code;
 
 import com.woowacamp.soolsool.global.exception.ErrorCode;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum MemberErrorCode implements ErrorCode {
-    
+
     INVALID_LENGTH_ADDRESS(400, "M101", "회원 주소는 100자보다 길 수 없습니다."),
     NO_CONTENT_ADDRESS(400, "M102", "회원 주소는 null이거나 공백일 수 없습니다."),
 
@@ -26,6 +26,9 @@ public enum MemberErrorCode implements ErrorCode {
 
     INVALID_LENGTH_PHONE_NUMBER(400, "M112", "회원 전화번호는 13자보다 길 수 없습니다."),
     NO_CONTENT_PHONE_NUMBER(400, "M113", "회원 전화번호는 null이거나 공백일 수 없습니다."),
+
+    MEMBER_NO_INFORMATION(400, "M101", "회원 정보를 찾을 수 없습니다."),
+    MEMBER_NO_ROLE_TYPE(400, "M101", "일치하는 회원 타입이 없습니다."),
     ;
 
     private final int status;
