+++ src/main/java/com/woowacamp/soolsool/global/exception/DefaultErrorCode.java
@@ -9,8 +9,6 @@ public enum DefaultErrorCode implements ErrorCode {
 
     DEFAULT_ERROR(400, "D101", "기본 예외입니다."),
     DEFAULT_VALIDATION_ERROR(400, "D102", "요청 정보 오류입니다."),
-    MEMBER_NO_INFORMATION(400, "M101", "회원 정보를 찾을 수 없습니다."),
-    MEMBER_NO_ROLE_TYPE(400, "M101", "일치하는 회원 타입이 없습니다."),
     ;
 
     private final int status;
